Uses proper units instead of workaround
The culprit was the .loopend = '1m' on BaseApp.js, which while it makes
sense for the first three sessions, in which we are basically looping
over a one measure phrase, here, in the 4th session, the samples are of
two measures, and therefore the Tone.Sequence callback was being
invoked to the beginning every time after one measure has passed.

It was a simple matter of pulling it "up" into the CircleApp.js class
and problem solved.